pr feedback initial pass
