+++ js/packages/mobile-wallet-adapter-walletlib/src/index.ts
@@ -1,3 +1,4 @@
 export * from './mwaSessionEvents.js';
 export * from './resolve.js';
 export * from './useMobileWalletAdapterSession.js';
+export * from './useDigitalAssetLinks.js';
